Move examples, output, and dev index under test/
Consolidate snapshot fixtures and the local example browser into test/ so the repo root stays focused on the library, and point Vite at test/ as its root.
