+++ test/examples/hello-world.js
