+++ test/examples/index.js
