+++ test/examples/option-cellsize.js
