+++ test/examples/option-dimensions.js
