+++ test/examples/option-padding.js
