+++ test/examples/option-style.js
