+++ test/examples/option-word.js
