+++ test/examples/render-content.js
