+++ vite.config.js
@@ -2,6 +2,7 @@ import {defineConfig} from "vitest/config";
 import path from "path";
 
 export default defineConfig({
+  root: path.resolve(__dirname, "test"),
   test: {
     environment: "jsdom",
   },
